Getting the change input port to work
